requirements.txt
# Data Manipulation
pandas==2.1.1
numpy==1.26.0
openpyxl==3.1.2

# Machine Learning & Stats
scikit-learn==1.3.1
scipy==1.11.3
statsmodels==0.14.0

# Gradient Boosting (The Heavy Hitters)
catboost==1.2.2
xgboost==2.0.0
lightgbm==4.1.0

# Visualization
matplotlib==3.8.0
seaborn==0.13.0
plotly==5.17.0

# Explanation & Utility
shap==0.42.1
joblib==1.3.2
python-dotenv==1.0.0
pytest==7.4.2
# Streamlit & Reporting
streamlit==1.28.0
streamlit-shap==1.0.1
fpdf2==2.7.4
